+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 20.08
 
+* Deprecate SSL configuration
 * New API to simulate a user pressing DTMF keys
 
   * `PUT /calls/{call_id}/dtmf`